Fix:댓글창 따로따로 댓글이 생기도록 수정
